scripts/requirements.txt

# FastAPI and server dependencies
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic>=1.10.7
python-multipart==0.0.6

# Flask (for backward compatibility if needed)
flask==3.0.3
flask-cors==5.0.0

# Machine learning dependencies
scikit-learn==1.5.2
numpy==2.1.3
joblib==1.4.2
xgboost==2.1.1

# PDF Processing and AI
google-generativeai==0.8.3
PyPDF2==3.0.1

# Environment variables
python-dotenv==1.0.0

# Utilities
requests==2.28.2
